+++ core/commons/docker_constants/docker_constants.go
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2021 - present Kurtosis Technologies LLC.
+ * All Rights Reserved.
+ */
+
+package docker_constants
+
+const (
+	// This is the domain name that Docker will give the host machine when inside a container
+	HostMachineDomainInsideContainer = "host.docker.internal"
+)